--- src/components/Layout/LayoutSubPages/Examination/ExaminationSubPages/Hallticket/Hallticket.css
@@ -1,45 +1,0 @@
-/* Style the entire table */
-table {
-    width: 100%; /* Make the table take up 100% of its container's width */
-    border-collapse: collapse; /* Collapse borders to remove spacing between cells */
-    margin-bottom: 20px; /* Add some spacing below the table */
-  }
-  
-  /* Style table headers (th) */
-  th {
-    background-color: #f2f2f2; /* Background color for header cells */
-    text-align: left; /* Align text to the left within header cells */
-    padding: 10px; /* Add padding to header cells for spacing */
-    border: 1px solid #ddd; /* Add a border around header cells */
-  }
-  
-  /* Style table rows (tr) */
-  tr {
-    border: 1px solid #ddd; /* Add a border around each row */
-  }
-  
-  /* Style table cells (td) */
-  td {
-    padding: 10px; /* Add padding to data cells for spacing */
-    border: 1px solid #ddd; /* Add a border around data cells */
-  }
-  
-  /* Alternating row colors for better readability */
-  tr:nth-child(even) {
-    background-color: #f2f2f2; /* Alternate row background color */
-  }
-  
-  /* Hover effect on table rows */
-  tr:hover {
-    background-color: #e0e0e0; /* Background color on hover */
-  }
-  
-  .setline1{
-    display: flex;
-    justify-content: space-around;
-    padding: 10px;
-  }
-  .setline1 {
-    text-align: left; /* Align the button to the left within the container */
-    padding-left: 10px; /* Adjust the left padding as needed */
-  }
